requirements.txt
django
psycopg2-binary
PyJWT
requests
channels
channels-redis
daphne